Add helper script to concatenate assets.
The script build_assets.sh can be used to concatenate all css/js assets from a given folder.
In addition it calculates the md5sum of the target file and puts the md5sum into the target file name. As a consequence, each time the assets are modified and reconcatenated, the md5sum and the file name also change. When the file name has changed, the browser doesn't automatically loads those assets from the browser instead of using the older version from the cache.