Add photos folder.
Because it might get tricky to store the captured photos at a folder outside the app's document root and still access it as static content, the photos will be stored at a folder called 'photos' in the document root.
There is a .gitignore file in there that keeps the folder tracked by git but ignores all captured images.