Add PHP script that return single config option as JSON.
This resolves #34.

Unfortunately, gphoto returns exit code 0 when the given option descriptor doesn't exist, and it doesn't return an error message, too.
As a consequence, the JSON file will contain an invalid, empty or null content if the option descriptor was invalid.
However, the GPhotoException content will be returned if one was thrown (such as when no camera is connected).